remedy return object of new 3p fit algorithm
When the initial code for the new divide and seek algorithm for 3p determination of mle models was prepared, the return object of the fit was a single row dataframe. However to be consistent with all previous return objects from mlefit this object had to be returned as a vector.